[ao] Added statistical threshold arg in Outlier Detector
Summary: The outlier detector has a feature where it's able to notify
the user if below the whole set of batches that passed through were used
in Outlier calculation, which mainly happens as a result of 0-errors.
This changes the code so that instead of comparing against a value like
30 as we were before, we now let the user pass in an optional fractional
value and if the ratio of the batches used was below that value, the
detector alerts the user.

Test Plan: python test/test_quantization.py TestFxDetectOutliers

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: d8cc802bb4df82325dccad03d3d879744c42527e
Pull Request resolved: https://github.com/pytorch/pytorch/pull/81174