Lower randint default dtype to the C++ API
`randint` is currently declared with `dtype=None` but the python
binding code is special-cased to default it to `long` instead. This
leads to an inconsistency between the eager-mode python API and
everything else. So, instead I move the default handling to C++.

Note the `_window` functions also need modified here because they call
`arange` and expect it to return a float by default.

ghstack-source-id: f84b68623b546c971d5abe5f27330072465ed781
Pull Request resolved: https://github.com/pytorch/pytorch/pull/81410