unflatten now has a free function version in torch.flatten in addition to the method in torch.Tensor.flatten.
Updated docs to reflect this and polished them a little.
For consistency, changed the signature of the int version of unflatten in
native_functions.yaml.

Some override tests were failing because unflatten has unusual
characteristics in terms of the .int and .Dimname versions having
different number of arguments so this required some changes
to test/test_override.py

ghstack-source-id: ba7603efff4a4eae760befbc63603cc2650b961f
Pull Request resolved: https://github.com/pytorch/pytorch/pull/81399

Trying to fix override test failures for unflatten