[ao] Standardize PerChannelDetector Output to be single level
Summary: Currently, the PerChannelDetector has a multi-layered output.

Example:
```
{'backend': 'qnnpack',
 'per_channel_status': {'block1.linear': {'per_channel_supported': True,
                                          'per_channel_used': False},
                        'block2.linear': {'per_channel_supported': True,
                                          'per_channel_used': False}}}
```

The issue with this is that when it comes to future features such as
visualizations where we need to go through this dictionary, it can be
hard because of the variable number of layers.

This changes the output format of the PerChannelDetector to have a
standard format.

Ex.)
```
module_fqn
|
--- feature name: val
```

Test Plan: python test/test_quantization.py TestFxModelReportDetector

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: d7a8956b817e98cd60bd88478014058f717329cf
Pull Request resolved: https://github.com/pytorch/pytorch/pull/81585