[ao] Add method in ModelReport to generate visualization
Summary: We created a ModelReportVisualization class, and the primary
way it is envisioned that it is accessed is:

```
model_report_visualization = model_reporter.generate_visualization()
```

This method only works after reports have been generated and it takes in
the generated reports and reformats them to be ordered by module, into
the format required by the ModelReportVisualization. It then generates
the visualization instance and returns that to the user.

Test Plan: python test/test_quantization.py TestFxModelReportClass.test_generate_visualization

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: cf23ac088b8ea49a22f5921c28d08d46b9727762
Pull Request resolved: https://github.com/pytorch/pytorch/pull/81589