Reduce the boilerplate to implement the getter of properties
We implement a template and we fill it up via CRTP. This heavily reduces
the ammount of repeated code.

Just testing the waters here. If you like this idea, I can easily extend
this idea to cover many of the properties that we currently implement.

N.b. It'd be nice to have proper `if constexpr` support for this one,
but here we are.

ghstack-source-id: 2bfa5a12f38d2e1b3488312db974bf1a2c6f93c6
Pull Request resolved: https://github.com/pytorch/pytorch/pull/81576